Bundle acorn and esotope.
The way 'npm install' works, if a parent package relies
on jalangi2, and that parent also depends on acorn or
esotope, the installation will not include a copy of those
packages under jalangi2's node_modules.  Since we rely on
those packages being there for instrumenting HTML, mark
them as bundled.